refactor(contact-form): co-locate component, merged unit tests, E2E in ContactForm/ folder
