--- __tests__/contact-rate-limit.test.ts
@@ -1,109 +1,0 @@
-// __tests__/contact-rate-limit.test.ts
-// Coverage gap (CG3 Task 3.4): the /api/contact rate-limit denial path.
-//
-// Behavioral test: the rate-limit factory is mocked to deny the request; the
-// POST handler must short-circuit with a 429, the standard error envelope
-// { ok: false, error: { code: 'rate_limited' } }, and an X-Request-Id header
-// — WITHOUT touching KV or Resend. Modeled on __tests__/contact-honeypot.test.ts.
-
-import { NextRequest } from 'next/server';
-import { beforeEach, describe, expect, it, vi } from 'vitest';
-
-const redisSetMock = vi.fn();
-const resendSendMock = vi.fn();
-const rateLimitMock = vi.fn();
-
-vi.mock('@/lib/rate-limit', () => ({
-  getClientIp: vi.fn(() => '127.0.0.1'),
-  getContactLimit: vi.fn(() => ({ limit: rateLimitMock })),
-  getRedis: vi.fn(() => ({ set: redisSetMock })),
-}));
-
-vi.mock('@/lib/ip-hash', () => ({
-  hashIp: vi.fn(async () => 'hashed-ip-test'),
-}));
-
-vi.mock('@/lib/log', () => ({
-  log: { info: vi.fn(), error: vi.fn(), warn: vi.fn(), debug: vi.fn() },
-}));
-
-vi.mock('resend', () => ({
-  Resend: class {
-    emails = { send: resendSendMock };
-  },
-}));
-
-function makeRequest(body: Record<string, unknown>): NextRequest {
-  return new NextRequest('http://localhost/api/contact', {
-    method: 'POST',
-    body: JSON.stringify(body),
-    headers: { 'Content-Type': 'application/json' },
-  });
-}
-
-const VALID_BODY = {
-  name: 'Real Name',
-  email: 'real@example.com',
-  message: 'A perfectly long-enough legitimate message',
-};
-
-describe('/api/contact — rate-limit denial path', () => {
-  beforeEach(() => {
-    process.env.RESEND_API_KEY = 'fake-key-for-tests';
-    vi.resetModules();
-    redisSetMock.mockReset();
-    resendSendMock.mockReset();
-    rateLimitMock.mockReset();
-  });
-
-  it('returns 429 with the rate_limited error envelope when the limit denies', async () => {
-    rateLimitMock.mockResolvedValueOnce({ success: false });
-    const { POST } = await import('@/app/api/contact/route');
-    const res = await POST(makeRequest(VALID_BODY));
-
-    expect(res.status).toBe(429);
-    const body = (await res.json()) as {
-      ok: false;
-      requestId: string;
-      error: { code: string; message: string };
-    };
-    expect(body.ok).toBe(false);
-    expect(body.error.code).toBe('rate_limited');
-    expect(typeof body.error.message).toBe('string');
-  });
-
-  it('sets the X-Request-Id header matching the envelope requestId', async () => {
-    rateLimitMock.mockResolvedValueOnce({ success: false });
-    const { POST } = await import('@/app/api/contact/route');
-    const res = await POST(makeRequest(VALID_BODY));
-
-    const body = (await res.json()) as { requestId: string };
-    expect(res.headers.get('x-request-id')).toBe(body.requestId);
-    expect(body.requestId).toMatch(/^[0-9a-f-]{36}$/);
-  });
-
-  it('does NOT persist to KV or call Resend when rate-limited', async () => {
-    rateLimitMock.mockResolvedValueOnce({ success: false });
-    const { POST } = await import('@/app/api/contact/route');
-    await POST(makeRequest(VALID_BODY));
-
-    // Rate-limit is the first gate — a denial must short-circuit before any
-    // downstream side effect (KV write, email send).
-    expect(redisSetMock).not.toHaveBeenCalled();
-    expect(resendSendMock).not.toHaveBeenCalled();
-  });
-
-  it('processes the submission normally when the limit allows', async () => {
-    rateLimitMock.mockResolvedValueOnce({ success: true });
-    redisSetMock.mockResolvedValueOnce('OK');
-    resendSendMock.mockResolvedValueOnce({ error: null });
-
-    const { POST } = await import('@/app/api/contact/route');
-    const res = await POST(makeRequest(VALID_BODY));
-
-    expect(res.status).toBe(200);
-    const body = (await res.json()) as { ok: boolean };
-    expect(body.ok).toBe(true);
-    expect(redisSetMock).toHaveBeenCalledOnce();
-  });
-});
