+++ components/client/ContactForm/index.ts
@@ -1,0 +1,1 @@
+export { ContactForm } from './ContactForm';
